fix:board like,unlike, schema 추가내용 반영

+++ src/main/java/com/sonhaenono/board/controller/BoardRestController.java
@@ -1,6 +1,8 @@
 package com.sonhaenono.board.controller;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.validation.Valid;
 
@@ -19,9 +21,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.sonhaenono.board.model.BoardDto;
+import com.sonhaenono.board.model.CommentDto;
 import com.sonhaenono.board.service.BoardService;
 import com.sonhaenono.exception.ApiException;
 import com.sonhaenono.exception.ExceptionEnum;
+import com.sonhaenono.util.SecurityUtil;
 
 
 @RestController
@@ -35,8 +39,8 @@ public class BoardRestController {
 	
 	@PostMapping
 	public ResponseEntity<?> writeArticle(@RequestBody @Valid BoardDto board) throws Exception {
-		// TODO: member 정보를 가져와서 board에 넣어주기
-		boardService.writeArticle(board);
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		boardService.writeArticle(memberId, board);
 		return new ResponseEntity<Boolean>(true, HttpStatus.OK);
 	}
 	@GetMapping
@@ -52,14 +56,48 @@ public class BoardRestController {
 		}
 		return new ResponseEntity<BoardDto>(article, HttpStatus.OK);
 	}
+	
+	@PostMapping("/{no}/like")
+	public ResponseEntity<?> ArticleLike(@PathVariable(name="no", required = true) int no) throws Exception {
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		//int like = BoardService.ArticleLike(no);
+		throw new ApiException(ExceptionEnum.API_NEW_FUNCTION_WAIT_REQUEST);
+	}
+	
+	@PostMapping("/{no}/unlike")
+	public ResponseEntity<?> articleUnlike(@PathVariable(name="no", required = true) int no) throws Exception {
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		//int like = BoardService.plusArticleLike(no);
+		throw new ApiException(ExceptionEnum.API_NEW_FUNCTION_WAIT_REQUEST);
+	}
+	
+	@PostMapping("/{boardNo}/comment")
+	public ResponseEntity<?> addComment(@PathVariable(name="boardNo") int boardNo, @RequestBody CommentDto comment) throws Exception {
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		
+		CommentDto savedComment = boardService.addComment(boardNo, memberId, comment);
+		
+		Map<String, Object> map = new HashMap<>();
+		return new ResponseEntity<Map<String, Object>>(map, HttpStatus.OK);
+
+	}
+	
 	@PutMapping("/{no}")
 	public ResponseEntity<?> updateArticle(@PathVariable(name="no", required = true) int no, @RequestBody BoardDto board) throws Exception {
-		boardService.updateArticle(no, board);
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		if(!boardService.isArticleOwner(no, memberId)) {
+			throw new ApiException(ExceptionEnum.ACCESS_DENIED_EXCEPTION);
+		}
+		boardService.updateArticle(no, memberId, board);
 		return new ResponseEntity<Boolean>(true, HttpStatus.OK);	
 	}
 	@DeleteMapping("/{no}")
 	public ResponseEntity<?> deleteArticle(@PathVariable(name="no", required = true) int no) throws Exception {
-		// TODO: 권한이 있는 사용자인지 체크 필요
+		String memberId = SecurityUtil.getCurrentUserId().get();
+		if(!boardService.isArticleOwner(no, memberId)) {
+			throw new ApiException(ExceptionEnum.ACCESS_DENIED_EXCEPTION);
+		}
+		
 		if(boardService.existArticle(no)) {
 			boardService.deleteArticle(no);
 			return new ResponseEntity<Boolean>(true, HttpStatus.OK);
